+++ utils/constants.js
@@ -1,6 +1,6 @@
 export const API = {
     URL: 'mongodb://localhost:27017/main',
-    PORT: 3007,
+    PORT: 3000,
 };
 
 const allowedOrigins = [`http://localhost:${API.PORT}`];
